requirements.txt
networkx
pytest-codestyle
pytest-cov
pytest-flakes
pytest-travis-fold
python-coveralls
PyYAML